extracting tag from code block
