+++ pages/components/buttons.md
@@ -1,0 +1,73 @@
+# Buttons
+
+## d2l-button
+
+The `d2l-button` element can be used just like the native button element, but also supports the `primary` attribute for denoting the primary button.
+
+### Example
+
+```html
+<script type="module">
+  import '@brightspace-ui/core/components/button/button.js';
+</script>
+<d2l-button>My Button</d2l-button>
+```
+
+### Accessibility
+
+To make your `d2l-button` accessible, use the following properties when applicable:
+
+| Attribute | Description |
+|--|--|
+| `aria-expanded` | [Indicate expansion state of a collapsible element](https://www.w3.org/WAI/PF/aria/states_and_properties#aria-expanded). Example: [d2l-more-less](https://github.com/BrightspaceUI/core/blob/f9f30d0975ee5a8479263a84541fc3b781e8830f/components/more-less/more-less.js#L158). |
+| `aria-haspopup` | [Indicate clicking the button opens a menu](https://www.w3.org/WAI/PF/aria/states_and_properties#aria-haspopup). Example: [d2l-dropdown](https://github.com/BrightspaceUI/core/blob/master/components/dropdown/dropdown-opener-mixin.js#L46). |
+| `description` | Use when text on button does not provide enough context. |
+
+## d2l-button-subtle
+
+The `d2l-button-subtle` element can be used just like the native `button`, but for advanced or de-emphasized actions.
+
+### Example
+
+```html
+<script type="module">
+  import '@brightspace-ui/core/components/button/button-subtle.js';
+</script>
+<d2l-button-subtle text="My Button" icon="tier1:gear"></d2l-button-subtle>
+```
+
+*Note:* It is strongly recommended to use `text` and `icon` as opposed to putting content in the `slot` to ensure that the recommended subtle button style is maintained.
+
+### Accessibility
+
+To make your `d2l-button-subtle` accessible, use the following properties when applicable:
+
+| Attribute | Description |
+|--|--|
+| `aria-expanded` | [Indicate expansion state of a collapsible element](https://www.w3.org/WAI/PF/aria/states_and_properties#aria-expanded). Example: [d2l-more-less](https://github.com/BrightspaceUI/core/blob/f9f30d0975ee5a8479263a84541fc3b781e8830f/components/more-less/more-less.js#L158). |
+| `aria-haspopup` | [Indicate clicking the button opens a menu](https://www.w3.org/WAI/PF/aria/states_and_properties#aria-haspopup). Example: [d2l-dropdown](https://github.com/BrightspaceUI/core/blob/master/components/dropdown/dropdown-opener-mixin.js#L46). |
+| `description` | Use when text on button does not provide enough context. |
+
+## d2l-button-icon
+
+The `d2l-button-icon` element can be used just like the native `button`, for instances where only an icon is displayed.
+
+### Example
+
+```html
+<script type="module">
+  import '@brightspace-ui/core/components/button/button-icon.js';
+</script>
+<d2l-button-icon text="My Button" icon="tier1:gear"></d2l-button-icon>
+```
+
+## Accessibility
+
+To make your `d2l-button-icon` accessible, use the following properties when applicable:
+
+| Attribute | Description |
+|--|--|
+| `aria-expanded` | [Indicate expansion state of a collapsible element](https://www.w3.org/WAI/PF/aria/states_and_properties#aria-expanded). Example: [d2l-more-less](https://github.com/BrightspaceUI/core/blob/f9f30d0975ee5a8479263a84541fc3b781e8830f/components/more-less/more-less.js#L158). |
+| `aria-haspopup` | [Indicate clicking the button opens a menu](https://www.w3.org/WAI/PF/aria/states_and_properties#aria-haspopup). Example: [d2l-dropdown](https://github.com/BrightspaceUI/core/blob/master/components/dropdown/dropdown-opener-mixin.js#L46). |
+| `text` | Acts as a primary label and tooltip and is **REQUIRED**. |
+| `aria-label` | Acts as a primary label. If `text` AND `aria-label` are provided, `aria-label` is used as the primary label, `text` is used as the tooltip. |
